% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{combinations.partial}
\alias{combinations.partial}
\title{Example list of combinations generated by comb_gen function.}
\description{
Generated with example sample set that contains 4 replicates each of heart
and muscle from GTeX. 5 combinations generated per replicate level.
}
