% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{combinations}
\alias{combinations}
\title{Example list of combinations generated by comb_gen function.}
\description{
Generated with example sample set that contains 10 replicates each of heart
and muscle from GTeX. 20 combinations generated at each replicate level.
}
