% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{condition_table.full}
\alias{condition_table.full}
\title{Example table of sample names and conditions}
\description{
Example condition table with two columns. Column 1 contains sample name,
column 2 contains condition name. Use in conjuncture with count_table.
Contains 10 samples each from heart and muscle.
}
