% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_table.R
\name{condition_table}
\alias{condition_table}
\title{Example table of sample names and conditions}
\description{
Example condition table with two columns. Column 1 contains sample name,
column 2 contains condition name. Use in conjuncture with count_table
}
