% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{count_table.full}
\alias{count_table.full}
\title{Example count table of GTEx RNA-seq experiment}
\description{
Count table generated using data downloaded from Recount2 website
with processed Raw GTEx dataset. Table has samples on columns and genes
on rows.
Contains 10 samples each from heart and muscle.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
