% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table.R
\name{count_table}
\alias{count_table}
\title{Example count table of GTEx RNA-seq experiment}
\description{
Count table generated using data downloaded from Recount2 website
that processed Raw GTEx dataset.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
