% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{deg}
\alias{deg}
\title{Example list of DE genes generated by edgeR}
\description{
Generated using object count_table.filtered and control='heart' as inputs to
erssa_edger function. Default parameters used. Contains list of list of
vectors. Each vector contains DE genes at a particular replicate
level generated by edgeR in a given sample combination.
}
