% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_gen.R
\name{comb_gen}
\alias{comb_gen}
\title{Calculate unique combinations of samples at varying replicate number}
\usage{
comb_gen(condition_table, n_repetition = 20, seed)
}
\arguments{
\item{condition_table}{A condition table with two columns and each sample as a row. Column 1 contains sample names and Column 2 contains sample condition (e.g. Control, Treatment).}

\item{n_repetition}{The number of count tables to generate at each replicate number. DEFAULT = 20.}

\item{seed}{An optional seed to generate reproducible random sampling. DEFAULT = No seed}
}
\value{
A list of character vectors containing sample combinations. Each element in the list corresponds to a replicate level. Each combination within the character vector is a single string with sample names separated by semicolon.
}
\description{
\code{comb_gen} function takes the list of samples in each condition
and calculate unique combination of sample names that allow
subsampling at varying replicate number.
}
\details{
At each replicate number, the total number of unique combination
of samples is computed. For example, 10 condition A samples subsampled
at replicate number of 2 has 45 unique combinations.

When the total number of possible combinations at a particular
replicate number is more than the specified x number of
repetition (default=20), then only x unique combinations are selected.

When the total number of possible combination is smaller than the
specified x number of repetitions, then only unique combinations
are selected. For example, 10 samples subsample for 9 replicates
has 10 unique combinations and only 10 combinations will be selected
instead of 20.

This is repeated for both conditions and another level of combination
is performed to combine samples from the two conditions. Again, only
x number of unique combinations selected. When total unique combination
is smaller than x, then all unique combinations are selected.

Selected combinations of samples at each replicate number are then
returned.
}
\examples{
combinations = comb_gen(condition_table, n_repetition=20)

}
