% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_filter.R
\name{count_filter}
\alias{count_filter}
\title{Filter RNA-seq expression table}
\usage{
count_filter(count_table, cutoff = 10)
}
\arguments{
\item{count_table}{Input RNA-seq count table.}

\item{cutoff}{Any gene with average count below this number will be filtered.}
}
\value{
The filtered count table is returned.
}
\description{
\code{count_filter} Function filters the RNA-seq count table to remove
low expression genes at an user-defined averaged count cutoff
(Default=10).
}
\examples{
count_filter(count_table, cutoff=10)

}
