% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_filter.R
\name{count_filter}
\alias{count_filter}
\title{Filter RNA-seq expression table}
\usage{
count_filter(count_table, cutoff = 1)
}
\arguments{
\item{count_table}{Input RNA-seq count table.}

\item{cutoff}{Any gene with average CPM below this cutoff value will be filtered out.}
}
\value{
The filtered count table is returned.
}
\description{
\code{count_filter} function filters the RNA-seq count table to remove
low expression genes at an defined average CPM cutoff
(Default=1).
}
\examples{
count_table.filtered = count_filter(count_table, cutoff=1)

}
