% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_filter.R
\name{count_filter}
\alias{count_filter}
\title{Filter RNA-seq expression table}
\usage{
count_filter(count_table = NULL, cutoff = 1, path = ".")
}
\arguments{
\item{count_table}{A RNA-seq count matrix with genes on each row and samples on each column.}

\item{cutoff}{Any gene with average CPM below this cutoff value will be filtered out.}

\item{path}{Path to save the filtered count table}
}
\value{
The filtered count table.
}
\description{
\code{count_filter} function filters the RNA-seq count table to remove
low expression genes at an defined average CPM cutoff
(Default=1).
}
\details{
The filtered count table is saved to the drive.
}
\examples{
#load simple GTeX example count table
data(count_table, package = "ERSSA")
count_table.filtered = count_filter(count_table)

}
\author{
Zixuan Shao, \email{Zixuanshao.zach@gmail.com}
}
