% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{count_table.filtered.partial}
\alias{count_table.filtered.partial}
\title{Example count table of GTEx RNA-seq experiment, fitered by count_filter
function}
\description{
A filtered version of the example count_table.partial.
Contains 4507 genes that passed the filter.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
