% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{count_table.filtered}
\alias{count_table.filtered}
\title{Example count table of GTEx RNA-seq experiment, fitered by count_filter function}
\description{
Count table is filtered version of example count_table. Contain 16766 genes that
pass the filter.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
