% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obj.R
\name{count_table}
\alias{count_table}
\title{Example count table of GTEx RNA-seq experiment}
\description{
Count table generated using data downloaded from Recount2 website
that processed Raw GTEx dataset. Table has samples on columns and genes
on rows.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
